+++ tag.go
@@ -7,6 +7,12 @@ import (
 	"sync"
 )
 
+const (
+	// when a tag is split on ',' if there are 2 elements,
+	// the second indicates the occurs count
+	occursIndicator = 2
+)
+
 var fieldRepCache sync.Map // map[reflect.Type]structRepresentation
 
 type structRepresentation struct {
@@ -23,7 +29,7 @@ type fieldRepresentation struct {
 func parseTag(tag string, prev int) (int, int, int, int, error) {
 	var occursSize int
 	ss := strings.Split(tag, ",")
-	if len(ss) == 2 {
+	if len(ss) == occursIndicator {
 		o, err := strconv.Atoi(ss[1])
 		if err != nil {
 			return 0, 0, 0, 0, err
